requirements.txt
beautifulsoup4==4.9.3
Flask==2.0.2
googlesearch_python==1.0.1
numpy==1.24.0
pandas==1.3.4
python_dateutil==2.8.2
requests==2.25.1
scikit_learn==1.0.1
whois==0.9.13
gunicorn==20.1.0
